OCP-1160: cleanup
OCP-1160: rename-folder-structure-to-use-kebab

OCP-1160: cleanup
